requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cups==0.0.6
Flask==3.0.2
Flask-Cors==4.0.0
gpip==0.4.7
gunicorn==21.2.0
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dotenv==1.0.1
razorpay==1.4.1
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1
